requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.2.2
mysql-connector-python==8.2.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0.post1
fastapi_utilities==0.1.3.1